Seperated legend and the tooltip
